feat: the menu is refactored
