+++ src/components/Layout/Layout.tsx
@@ -1,14 +1,16 @@
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
+import { XIcon, MenuIcon } from '@heroicons/react/outline'
 
 import { composeClasses } from 'dd360-ds/lib'
-import { Flex, Overflow } from 'dd360-ds'
+import { Flex, Overflow, Text } from 'dd360-ds'
 
 import { useTheme } from '@/pages/store/theme-store'
 import FloatingNav, { Entries } from './FloatingNav'
 import Footer from './Footer'
 import Navbar from './Navbar/Navbar'
-import SideBar from './SideBar'
+import SideBarDocs from './SideBarDocs'
+import CircleCustom from './Navbar/CircleCustom'
 
 function parseIdByName(name: string) {
   return name
@@ -21,7 +23,11 @@ function Layout({ children }: { children: JSX.Element }) {
   const router = useRouter()
   const showSidebar = router.pathname.startsWith('/docs/')
   const [entries, setEntries] = useState<Entries[]>([])
-  const { isLightTheme } = useTheme()
+  const {
+    isLightTheme,
+    themeObject: { extendedPalette }
+  } = useTheme()
+  const [isActiveButtonMobile, setIsActiveButtonMobile] = useState(false)
 
   useEffect(() => {
     const headers = document.getElementsByName('floating-nav')
@@ -53,20 +59,59 @@ function Layout({ children }: { children: JSX.Element }) {
         >
           <Navbar />
 
-          <Flex>
-            <div className="hidden md:flex">
-              <SideBar />
+          <Flex
+            alignItems="center"
+            gap="2"
+            className={composeClasses(
+              'py-2 px-4 md:hidden',
+              extendedPalette.cardBackground
+            )}
+          >
+            <Text className={extendedPalette.primaryText}>Menu</Text>
+            <CircleCustom
+              size="32px"
+              backgroundColor={extendedPalette.componentBgPrimaryHex}
+              onClick={() =>
+                setIsActiveButtonMobile &&
+                setIsActiveButtonMobile((prev) => !prev)
+              }
+            >
+              {isActiveButtonMobile ? (
+                <XIcon
+                  className="cursor-pointer"
+                  style={{ color: extendedPalette.navbarIconHex }}
+                  width={16}
+                />
+              ) : (
+                <MenuIcon
+                  className="cursor-pointer"
+                  style={{ color: extendedPalette.navbarIconHex }}
+                  width={16}
+                />
+              )}
+            </CircleCustom>
+          </Flex>
+          <Flex className="relative">
+            <div
+              className={composeClasses(
+                isActiveButtonMobile ? 'flex' : 'hidden',
+                'md:flex'
+              )}
+            >
+              <SideBarDocs />
             </div>
 
-            <Overflow
+            <div
               className="layout-content grid h-full px-8 md:px-16 m-auto"
               style={{ height: 'calc(100vh - 57px)' }}
             >
-              <article className="w-full max-w-full">{children}</article>
+              <article className="w-full max-w-full overflow-auto">
+                {children}
+              </article>
               <article className="hidden w-full max-w-[128px] mt-[76px] md:block">
                 <FloatingNav entries={entries} />
               </article>
-            </Overflow>
+            </div>
           </Flex>
           <Footer />
         </main>
