+++ src/components/Layout/Navbar/Navbar.tsx
@@ -1,13 +1,12 @@
-import { useCallback, useState } from 'react'
-import { useRouter } from 'next/router'
+import { useState } from 'react'
 import { useTranslation } from 'next-i18next'
 import Link from 'next/link'
 import {
+  XCircleIcon,
   MoonIcon,
   SunIcon,
   DesktopComputerIcon
 } from '@heroicons/react/outline'
-import { XCircleIcon } from '@heroicons/react/solid'
 
 import { Circle, Flex, Transition, Dropdown } from 'dd360-ds'
 import { composeClasses } from 'dd360-ds/lib'
@@ -21,24 +20,7 @@ import { Dd360Icon, GitHubIcon } from '../../icons'
 import SideBar from '../SideBar'
 import CircleCustom from './CircleCustom'
 import Search from './Search'
-
-const menu = [
-  {
-    key: 1,
-    label: 'components',
-    link: '/components'
-  },
-  {
-    key: 2,
-    label: 'docs',
-    link: '/docs/get-started/getting-started'
-  },
-  {
-    key: 3,
-    label: 'showcases',
-    link: '/showcases'
-  }
-]
+import MainLinks from './MainLinks'
 
 const getThemeIcon = (theme: string, width = 12, color = 'currentColor') => {
   if (theme === 'light') return <SunIcon width={width} color={color} />
@@ -49,69 +31,47 @@ const getThemeIcon = (theme: string, width = 12, color = 'currentColor') => {
 
 function Navbar() {
   const { t } = useTranslation('common')
-  const router = useRouter()
   const {
     themeObject: { extendedPalette },
     themeOption,
     onClickTheme
   } = useTheme()
-
   const [isActiveButtonMobile, setIsActiveButtonMobile] = useState(false)
-
-  const renderLinks = useCallback(() => {
-    return menu.map(({ key, label, link }) => (
-      <Link
-        key={key}
-        href={link}
-        locale={router?.locale}
-        className={composeClasses(
-          'h-full flex items-center hover:border-blue-500 font-medium',
-          router.pathname.includes(label)
-            ? 'text-blue-400'
-            : extendedPalette.navbarLink
-        )}
-      >
-        {t(label)}
-      </Link>
-    ))
-  }, [router, extendedPalette, t])
 
   return (
     <nav
       className={composeClasses(
-        'w-full sticky top-0 z-10 border-b',
-        extendedPalette.barBackground,
-        extendedPalette.sidebarBorder
+        'w-full sticky top-0 z-10 border-b flex flex-col-reverse md:flex-row',
+        extendedPalette.sidebarBorder,
+        isActiveButtonMobile
+          ? `${extendedPalette.barMobileBackground} sm:${extendedPalette.barBackground}`
+          : extendedPalette.barBackground
       )}
     >
       {isActiveButtonMobile && (
-        <Transition
-          animationStart="zoomInDown"
-          show={isActiveButtonMobile}
-          duration={200}
-          className="w-full bg-blue-50"
-        >
-          <div className="pt-4 pr-4 cursor-pointer absolute z-50 right-0">
-            <XCircleIcon
-              onClick={() => setIsActiveButtonMobile(!isActiveButtonMobile)}
-              width={20}
-            />
-          </div>
-          <SideBar />
-        </Transition>
+        <div className="md:hidden">
+          <Transition
+            animationStart="zoomInDown"
+            show={isActiveButtonMobile}
+            duration={200}
+            className="w-full"
+          >
+            <SideBar setIsActiveButtonMobile={setIsActiveButtonMobile} />
+          </Transition>
+        </div>
       )}
       <Flex
         justifyContent="between"
         alignItems="center"
         gap="4"
-        className="h-14 py-2 mx-8 px-4 lg:px-16 2xl:px-0 max-w-8xl flex-nowrap"
+        className="w-full h-14 py-2 mx-auto px-4 flex-nowrap"
       >
         <Link href="/">
           <Dd360Icon color={extendedPalette.logoColorHex} />
         </Link>
 
         <ul className="hidden items-center gap-8 h-12 md:flex">
-          {renderLinks()}
+          <MainLinks className="font-medium" />
         </ul>
 
         <Flex gap="2">
@@ -141,7 +101,7 @@ function Navbar() {
                   alignItems="center"
                   gap="1"
                   className={composeClasses(
-                    'p-2 text-xs',
+                    'p-2 text-xs cursor-pointer',
                     theme === themeOption &&
                       extendedPalette.componentBgSecondary
                   )}
@@ -164,10 +124,18 @@ function Navbar() {
             height="36px"
             backgroundColor="transparent"
           >
-            <MenuIcon
-              onClick={() => setIsActiveButtonMobile(!isActiveButtonMobile)}
-              color={extendedPalette.navbarIconHex}
-            />
+            {isActiveButtonMobile ? (
+              <XCircleIcon
+                width={24}
+                onClick={() => setIsActiveButtonMobile(!isActiveButtonMobile)}
+                color={extendedPalette.navbarIconHex}
+              />
+            ) : (
+              <MenuIcon
+                onClick={() => setIsActiveButtonMobile(!isActiveButtonMobile)}
+                color={extendedPalette.navbarIconHex}
+              />
+            )}
           </Circle>
         </Flex>
       </Flex>
